requirements.txt
django==4.2.7
django-extensions==3.2.3
requests==2.31.0
django-bootstrap5==23.3
gunicorn==21.2.0
mysqlclient==2.2.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
django-apscheduler==0.6.2